Separate export/prepare for Neutron data
In order to implement the provider router migration, it is easier to have the
intermediate data representing all the Neutron objects.  Separate the commands
to prepare and export Neutron data, where the output of the prepare command is
the input of the export command.  The output of the prepare command will be
used for other purposes in the future.

Change-Id: I620e10a54db0bbdb623258f02bc619ef8f31fd4f