Separate out Provider Router conversion
Define a new command, 'pr2er', which performs the conversion from the
provider router of v1.9.X to an edge router in v5.X.X, including uplink
networks.  The reason for the separation is that this operation
modifies the Neutron DB and it is safer to separate it from more
harmless 'migrate' commmand.

Change-Id: I4698fd510783edf4a17f628cbbe72282fe8d5922