Update antispoof port skip-checks
If a port refers to a chain which doesn't have a rules section
(i.e. if the rules list was empty int he prepare stage), then
skip any allowed_address_pair creation as we don't know the
intent of those rules' removals.

If a port is referring to a port which is not in neutron anymore,
don't try to update that port's allowed address pairs.

Add a list of skipped objects to the summary report. Added mock
plugin to write context and added anti-spoof test's port to the
list of mocked neutron ports.  This way the check for the missing
port has a list of ports to check against, and can see that the
specified port is missing.

Change-Id: Ie41b61e9820d0b917df82cea583f3f6c097f7099