Fix interfaces to allow for >1
The router_interfaces were set so the ID of the returned objects
was the router ID.  This would map all router's interfaces into
one entry int eh prep map, overwriting any other router interface
entries which may have come around (i.e. if a router has multiple
interfaces).

Now, instead, map the router_id to its own field and use that to
send to the add_Router_interface command, and then set the obj's
ID to the port ID, as a port can only be bound to a maximum of
1 interface, meaning that there should be no collisions.

Prior to sending the router_interface_add in the Writer, though,
we need to reset the ID of the sent object to the router ID,
as neutron expects the ID to be the device ID, rather than its
own unique ID (router interfaces are a special case).

Change-Id: I425f477d55c4cf4be800e882a6cd7f90e58dcf93